backend/requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==24.2.0
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-script==1.0.1
djangorestframework==3.15.2
djangorestframework-gis==1.1
fiona==1.9.2
geopandas==0.13.2
gunicorn==23.0.0
idna==3.10
importlib-metadata==8.5.0
numpy==1.24.4
packaging==24.1
pandas==2.0.3
psycopg2==2.9.10
pyproj==3.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.2
requests==2.32.3
shapely==2.0.6
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
zipp==3.20.2
